Restructure Project
Creation of one repository to house all project files. Movemented of all
non-code related documents to docs folder, and creaton of app folder to
house main application components.

Use of Flask Blueprints to segment applicaton into local and web
components, as well as, creation of the class folder to house controller
and handler classes defined in the project docs.